requirements.txt
numpy
tensorflow
portaudio
python-sounddevice
kivy